requirements.txt
icecream==2.1.3
requests==2.31.0
pytest==8.1.0
pytest-mock==3.12.0
python-decouple==3.8


